refactor(autolabcli-ng): Change project structure.
