+++ index.js
@@ -1,12 +1,5 @@
 #!/usr/bin/env node
 
-let program = require('caporal');
+const controller = require('./lib/controller');
 
-const init = require('./lib/cli/init');
-
-program
-  .version('1.0.0')
-  .description('A Command Line Interface (CLI) for Autolab');
-init.add(program);
-
-program.parse(process.argv);
+controller.start();
