+++ lib/cli/output/init.js
@@ -1,0 +1,5 @@
+module.exports = {
+  send: (output, logger) => {
+    logger.info(output);
+  }
+}
